Another Refactor
Cleaned up Arceus a bunch more.

+++ lib/commands/trim.dart
@@ -4,7 +4,6 @@ import 'package:arceus/extensions.dart';
 import 'package:args/command_runner.dart';
 import 'package:arceus/arceus.dart';
 import 'package:cli_spin/cli_spin.dart';
-import 'package:interact/interact.dart';
 
 class TrimCommand extends Command {
   @override
@@ -27,36 +26,24 @@ class TrimConstCommand extends Command {
 
   TrimConstCommand() {
     argParser.addOption("const",
-        abbr: "c", help: "The name of the constellation you wish to delete.");
+        abbr: "c",
+        help: "The name of the constellation you wish to delete.",
+        mandatory: true);
   }
 
   @override
   Future<void> run() async {
-    String constellationName;
-    if (argResults!.option("const") == null) {
-      final consts =
-          Directory(Arceus.constFolderPath).listSync(recursive: true);
-      if (consts.isEmpty) {
-        throw Exception("No constellations found.");
-      }
-      final selection = Select(
-          prompt: "Select constellation to delete.",
-          options: [
-            ...consts.map((e) => e.path.getFilename(withExtension: false)),
-            "Cancel"
-          ]).interact();
-      if (selection == consts.length) {
-        return;
-      }
-      constellationName =
-          consts[selection].path.getFilename(withExtension: false);
-    } else {
-      constellationName = argResults!.option("const")!;
+    String constellationName = findOption("const");
+    final kit = File("${Arceus.constFolderPath}/$constellationName.skit");
+    if (!await kit.exists()) {
+      throw FileSystemException(
+          "Cannot find skit with the name of $constellationName.",
+          "${Arceus.constFolderPath}/$constellationName.skit");
     }
     final spinner = CliSpin(
             text: "Deleting $constellationName...", spinner: CliSpinners.moon)
         .start();
-    final kit = File("${Arceus.constFolderPath}/$constellationName.skit");
+
     await kit.delete();
     spinner.success("Deleted $constellationName!");
     return;
